Add files via upload
I am having trouble getting the image to correctly appear and the colors of the text aren't applying. The text with the different fonts are not applying correctly, some text with the same point size are appearing as different sizes. I have been working on the about page, I am not sure what page we are starting on.